+++ dmd/dmdbase.py
@@ -2,7 +2,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 
 
-class DMD(object):
+class DMDBase(object):
 	"""
 	Dynamic Mode Decomposition
 
@@ -11,8 +11,11 @@ class DMD(object):
 	:param numpy.ndarray X: the input matrix with dimension `m`x`n`
 	:param int k: rank truncation in SVD
 	"""
-	def __init__(self, k=None):
-		self.k = k
+	def __init__(self, svd_rank=0, tlsq_rank=0):
+
+		self.svd_rank  = svd_rank
+		self.tlsq_rank = tlsq_rank
+
 		self._basis = None	# spatial basis vectors
 		self._mode_coeffs = None
 		self._eigs = None
@@ -20,6 +23,7 @@ class DMD(object):
 		self._modes = None	# Phi
 		self._amplitudes = None	 # B
 		self._vander = None	 # Vander
+
 
 	@property
 	def modes(self):
@@ -41,47 +45,6 @@ class DMD(object):
 	def reconstructed_data(self):
 		return self._modes.dot(self._amplitudes).dot(self._vander)
 
-	def fit(self, X, Y=None):
-		"""
-		"""
-		n_samples = X.shape[1]
-		# split the data
-		if Y is None:
-			Y = X[:, 1:]
-			X = X[:, :-1]
-
-		#---------------------------------------------------------------------------
-		# Singular Value Decomposition
-		#---------------------------------------------------------------------------
-		U, s, V = np.linalg.svd(X, full_matrices=False)
-		V = np.conjugate(V.T)
-
-		if self.k is not None:
-			U = U[:, 0:self.k]
-			V = V[:, 0:self.k]
-			s = s[0:self.k]
-
-		Sinverse = np.diag(1. / s)
-
-		#---------------------------------------------------------------------------
-		# DMD Modes
-		#---------------------------------------------------------------------------
-		self._Atilde = np.transpose(U).dot(Y).dot(V).dot(Sinverse)
-		# exact DMD
-		self._basis = Y.dot(V).dot(Sinverse)
-		self._eigs, self._mode_coeffs = np.linalg.eig(self._Atilde)
-
-		self._modes = self._basis.dot(self._mode_coeffs)
-
-		#---------------------------------------------------------------------------
-		# DMD Amplitudes and Dynamics
-		#---------------------------------------------------------------------------
-		b = np.linalg.lstsq(self._modes, X[:, 0])[0]
-		self._amplitudes = np.diag(b)
-
-		self._vander = np.fliplr(np.vander(self._eigs, N=n_samples))
-
-		return self
 
 	def plot_eigs(self, show_axes=False, show_unit_circle=False):
 		"""
