+++ tests/test_dmdexact.py
@@ -1,5 +1,5 @@
 from unittest import TestCase
-from dmd.dmd import DMD
+from dmd.dmdexact import DMDExact
 
 import numpy as np
 
@@ -12,47 +12,47 @@ sample_data = np.load('tests/test_datasets/input_sample.npy')
 
 class TestDmd(TestCase):
 	def test_dmd_shape(self):
-		dmd = DMD()
+		dmd = DMDExact()
 		dmd.fit(X=sample_data)
 		assert dmd.modes.shape[1] == sample_data.shape[1] - 1
 
 	def test_dmd_truncation_shape(self):
-		dmd = DMD(k=3)
+		dmd = DMDExact(svd_rank=3)
 		dmd.fit(X=sample_data)
 		assert dmd.modes.shape[1] == 3
 
 	def test_dmd_amplitudes_1(self):
-		dmd = DMD()
+		dmd = DMDExact()
 		dmd.fit(X=sample_data)
 		assert dmd.amplitudes.shape == (14, 14)
 
 	def test_dmd_amplitudes_2(self):
-		dmd = DMD(k=3)
+		dmd = DMDExact(svd_rank=3)
 		dmd.fit(X=sample_data)
-		assert dmd.amplitudes.shape == (dmd.k, dmd.k)
+		assert dmd.amplitudes.shape == (dmd.svd_rank, dmd.svd_rank)
 
 	def test_dmd_vander(self):
-		dmd = DMD()
+		dmd = DMDExact()
 		dmd.fit(X=sample_data)
 		assert dmd.vander.shape == (14, 15)
 
 	def test_dmd_eigs_1(self):
-		dmd = DMD()
+		dmd = DMDExact()
 		dmd.fit(X=sample_data)
 		assert len(dmd.eigs) == 14
 
 	def test_dmd_eigs_2(self):
-		dmd = DMD(k=5)
+		dmd = DMDExact(svd_rank=5)
 		dmd.fit(X=sample_data)
 		assert len(dmd.eigs) == 5
 
 	def test_dmd_reconstructed_data(self):
-		dmd = DMD()
+		dmd = DMDExact()
 		dmd.fit(X=sample_data)
 		dmd_data = dmd.reconstructed_data
 		assert np.allclose(dmd_data, sample_data)
 
 	def test_dmd_plot_eigs(self):
-		dmd = DMD()
+		dmd = DMDExact()
 		dmd.fit(X=sample_data)
 		dmd.plot_eigs(show_axes=True, show_unit_circle=True)
